requirements.txt
mysqlclient
django
django-environ
djangorestframework
